backend/requirements.txt
pytesseract
opencv-python-headless
pillow
numpy
pdf2image
spacy
minio
python-dotenv
mysql-connector-python
PyYAML
boto3